apps/agent/scraper/requirements.txt
requests
beautifulsoup4
feedparser
pydantic
python-dotenv
tenacity
